:art: Move API work to separate files
Cleans up work and makes it unit testable
