+++ lib/api.js
@@ -1,0 +1,5 @@
+'use strict';
+
+const apis = require('./api/apis');
+
+module.exports = apis;
