+++ lib/utils.js
@@ -2,7 +2,6 @@
 
 const util = require('util');
 const moment = require('moment-timezone');
-const schemas = require('./database/schemas');
 
 moment.tz.setDefault('UTC');
 
@@ -210,165 +209,7 @@ const format = body => {
   return data;
 };
 
-/*
- * Formats raw live data in to consumable returns
- *
- * @param {array} content - Array of objects representing the raw DB query of content
- *
- * @returns {array} - formatted object of content
- */
-const formatAPI = (content, attrs) => {
-  return content.map(item => {
-    const result = {
-      id: item.id,
-      type: item.type,
-      type_slug: item['type-slug'], // eslint-disable-line camelcase
-      key: item.key,
-      key_slug: item['key-slug'], // eslint-disable-line camelcase
-    };
-
-    if (attrs) {
-      result.attributes = item.attributes;
-    }
-    else {
-      result.meta = {
-        url: `/api/types/${item['type-slug']}/${item.id}`,
-      };
-    }
-
-    return result;
-  });
-};
-
-/*
- * Determines raw live data organization
- *
- * @param {object} params - Request parameters
- *
- * @returns {object} - Object with sort key, sort direction, pagination count and offset
- */
-const liveSchema = singleItem('name', 'live', schemas).fields.map(field => {
-  return field.name;
-});
-
-const organizeAPI = (params, lookup) => {
-  let sort = params.sort || 'key';
-  let sortDir = params.sort_dir || 'asc';
-  let sortKeys = lookup || liveSchema;
-
-  let page = parseInt(params.page, 10) || 1;
-  let perPage = parseInt(params.per_page, 10) || 30;
-  let count = 0;
-
-
-  //////////////////////////////
-  // Sorting
-  //////////////////////////////
-  // Ensure sort is an available key
-  sortKeys = sortKeys.filter(key => {
-    if (key === 'attributes') {
-      return false;
-    }
-
-    return true;
-  });
-  if (sortKeys.indexOf(sort) < 0) {
-    if (lookup) {
-      sort = lookup[0];
-    }
-    else {
-      sort = 'key';
-    }
-  }
-
-  // Ensure sorting is either ascending or descending
-  if (sortDir !== 'asc' && sortDir !== 'desc') {
-    sortDir = 'asc';
-  }
-
-  //////////////////////////////
-  // Pagination
-  //////////////////////////////
-  // Ensure there is at least one page returned
-  if (page < 1) {
-    page = 1;
-  }
-
-  // Ensure pagination fits within parameters
-  if (perPage < 1) {
-    perPage = 1;
-  }
-  else if (perPage > 100) {
-    perPage = 100;
-  }
-
-  // Calculate offset
-  count = (page - 1) * perPage;
-
-  return {
-    sort: {
-      by: sort,
-      dir: sortDir,
-    },
-    page: {
-      offset: count,
-      limit: perPage,
-      page,
-    },
-  };
-};
-
-/*
- * Build Pagination Links
- *
- * @param {string} base - Base URL
- * @param {object} organize - Organized sort and pagination
- * @param {number} count - Total number of items
- */
-const pageAPI = (base, organize, count) => {
-  const baseURL = `/${base}?sort=${organize.sort.by}&sort_dir=${organize.sort.dir}&per_page=${organize.page.limit}`;
-  const pages = {
-    first: `${baseURL}&page=1`,
-  };
-  let total = Math.floor(count / organize.page.limit);
-  if (count % organize.page.limit !== 0) {
-    total += 1;
-  }
-
-  // If not the first page, add a previous link
-  if (organize.page.page !== 1) {
-    pages.prev = `${baseURL}&page=${organize.page.page - 1}`;
-  }
-  else {
-    pages.prev = false;
-  }
-
-  // Add next link if not on the last page
-  if (organize.page.page + 1 <= total) {
-    pages.next = `${baseURL}&page=${organize.page.page + 1}`;
-  }
-  else {
-    pages.next = false;
-  }
-
-  // Add last page
-  pages.last = `${baseURL}&page=${total}`;
-
-  // If first and last are the same, set them to false
-  if (pages.first === pages.last) {
-    pages.first = false;
-    pages.last = false;
-  }
-
-  return pages;
-};
-
 module.exports = {
-  api: { // TODO: Tests
-    format: formatAPI,
-    organize: organizeAPI,
-    pagination: pageAPI,
-  },
   singleItem,
   log,
   time,
